fix(builder): always-mount CalcStateProvider; lazy-load @smogon/calc engine
The conditional needsCalc ? <CalcStateProviderDynamic> : body rendered the body as a child of a dynamic(ssr:false) provider that returns null until its chunk loads. Two bugs: (1) the whole workspaceBody subtree unmounted/remounted when needsCalc first flipped true, resetting in-progress editor state; (2) synchronous getByTestId('speed-tiers-panel') failed because the panel never rendered synchronously.

Always-mount CalcStateProvider (static import) so children render immediately and never remount. Preserve the initial-bundle win by moving all @smogon/calc code into new calc/calc-engine.ts, loaded lazily inside use-calc-state.ts via a dynamic import() gated on calcEnabled; compute returns null outputs until the chunk resolves. Also reopen the import dialog on ?action=import for an already-mounted workspace. Fixes CI speed-tiers failure + remount + importOpen review findings.
