+++ apps/web/src/components/team-builder/calc/calc-state-context.tsx
@@ -6,12 +6,12 @@
  * working unchanged WITHOUT pulling in the engine.
  *
  * IMPORTANT: this barrel must NOT re-export `CalcStateProvider`. The provider
- * lives in ./calc-state-provider, which imports ../use-calc-state →
- * @smogon/calc. A static re-export here would drag the engine into every
- * always-on consumer that imports a hook from this barrel (dev bundlers don't
- * tree-shake the unused re-export), defeating the code-split. The provider is
- * imported directly: lazily via next/dynamic in team-workspace.tsx, and
- * directly from ./calc-state-provider in tests.
+ * lives in ./calc-state-provider, which imports ../use-calc-state. use-calc-state
+ * dynamically imports ./calc-engine → @smogon/calc at runtime. A static
+ * re-export here would drag the engine into every always-on consumer that
+ * imports a hook from this barrel (dev bundlers don't tree-shake unused
+ * re-exports), defeating the code-split. The provider is imported directly
+ * from ./calc-state-provider in team-workspace.tsx and in tests.
  */
 
 export {
